fix(config): cast openai moderation env values to int

+++ config/services.php
@@ -43,12 +43,8 @@ return [
 
     'openai' => [
         'key' => env('OPENAI_API_KEY'),
-        // Maximum characters to send to the moderation endpoint when composing
-        // title, author and description. Configurable via env.
-        'moderation_input_max' => env('OPENAI_MODERATION_INPUT_MAX', 3000),
-        // Cache TTL (in seconds) for moderation verdicts. Default: 1 day.
-        'moderation_cache_ttl' => env('OPENAI_MODERATION_CACHE_TTL', 86400),
-        // Failure cache TTL (in minutes) to avoid retry stampede when API fails.
-        'moderation_failure_cache_minutes' => env('OPENAI_MODERATION_FAILURE_CACHE_MINUTES', 2),
+            'moderation_input_max' => (int) env('OPENAI_MODERATION_INPUT_MAX', 3000),
+            'moderation_cache_ttl' => (int) env('OPENAI_MODERATION_CACHE_TTL', 86400),
+            'moderation_failure_cache_minutes' => (int) env('OPENAI_MODERATION_FAILURE_CACHE_MINUTES', 2),
     ],
 ];
